Add `aggregator` example.
This example connects to multiple sources of VBus data and starts
logging the incoming data. It also configures a timer to periodically
trigger the storage of the received data into a permanent storage.

This example focuses on data streams conforming to protocol version 1.0
or what the VBus specification calls: "packets". Packets are sent
regularly from the VBus master without the need to request them. So this
example can simply start listening for incoming data, decode it and
store the contained information.